+++ packages/react/src/atomic/Icon/src/utils/gradient/colorValidation/colorValidation.definitions.ts
@@ -1,6 +1,6 @@
 /**
  * Collection of known gradient color values
  */
-const GRADIENT_COLORS = new Set(["ai-interactive"]);
+const GRADIENT_COLORS = new Set(["ai-generative"]);
 
 export { GRADIENT_COLORS };